+++ packages/runtime/container-runtime/src/test/opLifecycle/outbox.spec.ts
@@ -17,6 +17,7 @@ import {
 	ISequencedDocumentMessage,
 } from "@fluidframework/driver-definitions/internal";
 import { MockLogger } from "@fluidframework/telemetry-utils/internal";
+import Sinon from "sinon";
 
 import type { ICompressionRuntimeOptions } from "../../compressionDefinitions.js";
 import { CompressionAlgorithms } from "../../compressionDefinitions.js";
@@ -242,9 +243,7 @@ describe("Outbox", () => {
 		opGroupingConfig?: OpGroupingManagerConfig;
 		immediateMode?: boolean;
 		flushPartialBatches?: boolean;
-		generateIdAllocationOp?: (
-			resubmitOutstandingRanges: boolean,
-		) => LocalBatchMessage | undefined; // Add this optional parameter for tests that need specific behavior
+		generateIdAllocationOp?: () => LocalBatchMessage | undefined;
 	}) => {
 		const { submitFn, submitBatchFn, deltaManager } = params.context;
 
@@ -277,7 +276,7 @@ describe("Outbox", () => {
 				state.opsResubmitted++;
 			},
 			opReentrancy: () => state.isReentrant,
-			generateIdAllocationOp: params.generateIdAllocationOp ?? (() => undefined), // Add the mock function here
+			generateIdAllocationOp: params.generateIdAllocationOp ?? (() => undefined),
 		});
 	};
 
@@ -332,16 +331,11 @@ describe("Outbox", () => {
 	});
 
 	it("Sending batches", () => {
-		// Store the ID allocation op to be returned by the mock
-		let idAllocOp: LocalBatchMessage | undefined;
+		// Create a Sinon stub for generateIdAllocationOp
+		const generateIdAllocationOpStub = Sinon.stub();
 		const outbox = getOutbox({
 			context: getMockContext(),
-			// Provide the mock generateIdAllocationOp
-			generateIdAllocationOp: () => {
-				const op1 = idAllocOp;
-				idAllocOp = undefined; // Consume the op
-				return op1;
-			},
+			generateIdAllocationOp: generateIdAllocationOpStub, // Pass the stub
 		});
 		const messages = [
 			createMessage(ContainerMessageType.FluidDataStoreOp, "0"),
@@ -355,16 +349,15 @@ describe("Outbox", () => {
 		// Flush 1
 		outbox.submit(messages[0]);
 		outbox.submit(messages[1]);
-		// No longer call submitIdAllocation directly
-		// outbox.submitIdAllocation(messages[2]);
-		// outbox.submitIdAllocation(messages[3]);
-		idAllocOp = messages[2]; // Set the op to be generated by the mock
-		outbox.flush(); // It should be included in this flush
+		// Configure the stub to return the ID alloc op on the first call during flush
+		generateIdAllocationOpStub.onCall(0).returns(messages[2]);
+		outbox.flush(); // Stub is called here
 
 		// Flush 2
 		outbox.submit(messages[4]);
-		// No ID alloc op for this flush
-		outbox.flush();
+		// Configure the stub to return undefined for the second call (no ID alloc)
+		generateIdAllocationOpStub.onCall(1).returns(undefined); //* Is this necessary?
+		outbox.flush(); // Stub is called here again
 
 		// Not Flushed
 		outbox.submit(messages[5]);
@@ -442,44 +435,41 @@ describe("Outbox", () => {
 	});
 
 	it("Batch ID added when applicable (ungrouped batch)", () => {
-		// Store the ID allocation op to be returned by the mock
-		let idAllocOp: LocalBatchMessage | undefined;
+		// Create a Sinon stub for generateIdAllocationOp
+		const generateIdAllocationOpStub = Sinon.stub<[], LocalBatchMessage | undefined>();
 		const outbox = getOutbox({
 			context: getMockContext(),
 			opGroupingConfig: {
 				groupedBatchingEnabled: false,
 			},
-			// Provide the mock generateIdAllocationOp
-			generateIdAllocationOp: () => {
-				const op1 = idAllocOp;
-				idAllocOp = undefined; // Consume the op
-				return op1;
-			},
+			generateIdAllocationOp: generateIdAllocationOpStub, // Pass the stub
 		});
 		const idAllocMessage = createMessage(ContainerMessageType.IdAllocation, "0");
 		// Flush 1 - resubmit multi-message batch including ID Allocation
-		// No longer call submitIdAllocation directly
-		// outbox.submitIdAllocation(createMessage(ContainerMessageType.IdAllocation, "0")); // Separate batch, batch ID not used
 		outbox.submit(createMessage(ContainerMessageType.FluidDataStoreOp, "1"));
 		outbox.submit(createMessage(ContainerMessageType.FluidDataStoreOp, "2"));
-		idAllocOp = idAllocMessage; // Set the op to be generated by the mock
+		// Configure stub for first flush
+		generateIdAllocationOpStub.onCall(0).returns(idAllocMessage);
 		outbox.flush("batchId-A");
 
 		// Flush 2 - resubmit single-message batch
 		outbox.submit(createMessage(ContainerMessageType.FluidDataStoreOp, "3"));
-		// No ID alloc op for this flush
+		// Configure stub for second flush (no ID alloc)
+		generateIdAllocationOpStub.onCall(1).returns(undefined);
 		outbox.flush("batchId-B");
 
 		// Flush 3 - resubmit blob attach batch
 		outbox.submitBlobAttach(createMessage(ContainerMessageType.BlobAttach, "4"));
 		outbox.submitBlobAttach(createMessage(ContainerMessageType.BlobAttach, "5"));
 		currentSeqNumbers.referenceSequenceNumber = 0;
-		// No ID alloc op for this flush
+		// Configure stub for third flush (no ID alloc)
+		generateIdAllocationOpStub.onCall(2).returns(undefined);
 		outbox.flush("batchId-C");
 
 		// Flush 4 - no batch ID given
 		outbox.submit(createMessage(ContainerMessageType.FluidDataStoreOp, "6"));
-		// No ID alloc op for this flush
+		// Configure stub for fourth flush (no ID alloc)
+		generateIdAllocationOpStub.onCall(3).returns(undefined);
 		outbox.flush(); // Ignored - No batchID given (not resubmit)
 
 		// Not Flushed (will not appear in batchesSubmitted or pendingOpContents)
@@ -558,16 +548,11 @@ describe("Outbox", () => {
 	});
 
 	it("Uses legacy path for legacy contexts", () => {
-		// Store the ID allocation op to be returned by the mock
-		let idAllocOp: LocalBatchMessage | undefined;
+		// Create a Sinon stub for generateIdAllocationOp
+		const generateIdAllocationOpStub = Sinon.stub<[], LocalBatchMessage | undefined>();
 		const outbox = getOutbox({
 			context: getMockLegacyContext() as IContainerContext,
-			// Provide the mock generateIdAllocationOp
-			generateIdAllocationOp: () => {
-				const op1 = idAllocOp;
-				idAllocOp = undefined; // Consume the op
-				return op1;
-			},
+			generateIdAllocationOp: generateIdAllocationOpStub, // Pass the stub
 		});
 		const messages = [
 			createMessage(ContainerMessageType.FluidDataStoreOp, "0"),
@@ -580,15 +565,15 @@ describe("Outbox", () => {
 		// Flush 1
 		outbox.submit(messages[0]);
 		outbox.submit(messages[1]);
-		// No longer call submitIdAllocation directly
-		// outbox.submitIdAllocation(messages[2]);
 		outbox.submit(messages[3]);
-		idAllocOp = messages[2]; // Set the op to be generated by the mock
+		// Configure stub for first flush
+		generateIdAllocationOpStub.onCall(0).returns(messages[2]);
 		outbox.flush();
 
 		// Flush 2
 		outbox.submit(messages[4]);
-		// No ID alloc op for this flush
+		// Configure stub for second flush (no ID alloc)
+		generateIdAllocationOpStub.onCall(1).returns(undefined);
 		outbox.flush();
 
 		// Legacy path submits individually.
@@ -598,17 +583,16 @@ describe("Outbox", () => {
 		// Flush calls: 1 for ID Alloc + main (Flush 1), 1 for Flush 2 main = 3
 		assert.equal(state.deltaManagerFlushCalls, 2);
 
-		//* Expecting this to be wrong 👇
 		// Note the expected CSN here is fixed to the batch's starting CSN
-		// Flush 1: ID Alloc [2] gets CSN 1. Main ops [0, 1, 3] get CSN 2.
-		// Flush 2: Main op [4] gets CSN 5 (1 + 3 + 1)
+		// Flush 1: ID Alloc [2] + Main ops [0, 1, 3] get CSN 1.
+		// Flush 2: Main op [4] gets CSN 5
 		const expectedMessageOrderWithCsn = [
 			// Flush 1 (ID Allocation generated)
 			[messages[2], 1],
 			// Flush 1 (Main)
-			[messages[0], 2],
-			[messages[1], 2],
-			[messages[3], 2],
+			[messages[0], 1],
+			[messages[1], 1],
+			[messages[3], 1],
 			// Flush 2 (Main)
 			[messages[4], 5],
 		] as const;
@@ -625,8 +609,8 @@ describe("Outbox", () => {
 	});
 
 	it("Compress if compression and grouping are enabled", () => {
-		// Store the ID allocation op to be returned by the mock
-		let idAllocOp: LocalBatchMessage | undefined;
+		// Create a Sinon stub for generateIdAllocationOp
+		const generateIdAllocationOpStub = Sinon.stub<[], LocalBatchMessage | undefined>();
 		const outbox = getOutbox({
 			context: getMockContext(),
 			compressionOptions: {
@@ -636,12 +620,7 @@ describe("Outbox", () => {
 			opGroupingConfig: {
 				groupedBatchingEnabled: true,
 			},
-			// Provide the mock generateIdAllocationOp
-			generateIdAllocationOp: () => {
-				const op1 = idAllocOp;
-				idAllocOp = undefined; // Consume the op
-				return op1;
-			},
+			generateIdAllocationOp: generateIdAllocationOpStub, // Pass the stub
 		});
 
 		const messages = [
@@ -653,11 +632,9 @@ describe("Outbox", () => {
 
 		outbox.submit(messages[0]);
 		outbox.submit(messages[1]);
-		// No longer call submitIdAllocation directly
-		// outbox.submitIdAllocation(messages[2]);
 		outbox.submit(messages[3]);
-		idAllocOp = messages[2]; // Set the op to be generated by the mock
-
+		// Configure stub for the flush
+		generateIdAllocationOpStub.onCall(0).returns(messages[2]);
 		outbox.flush();
 
 		// ID Alloc op [2] is generated and prepended to the main batch [0, 1, 3]
@@ -708,8 +685,8 @@ describe("Outbox", () => {
 	});
 
 	it("Does not compress if the batch is smaller than the configured limit", () => {
-		// Store the ID allocation op to be returned by the mock
-		let idAllocOp: LocalBatchMessage | undefined;
+		// Create a Sinon stub for generateIdAllocationOp
+		const generateIdAllocationOpStub = Sinon.stub<[], LocalBatchMessage | undefined>();
 		const outbox = getOutbox({
 			context: getMockContext(),
 			maxBatchSize: 1024,
@@ -720,12 +697,7 @@ describe("Outbox", () => {
 				minimumBatchSizeInBytes: 512,
 				compressionAlgorithm: CompressionAlgorithms.lz4,
 			},
-			// Provide the mock generateIdAllocationOp
-			generateIdAllocationOp: () => {
-				const op1 = idAllocOp;
-				idAllocOp = undefined; // Consume the op
-				return op1;
-			},
+			generateIdAllocationOp: generateIdAllocationOpStub, // Pass the stub
 		});
 
 		const messages = [
@@ -737,11 +709,9 @@ describe("Outbox", () => {
 
 		outbox.submit(messages[0]);
 		outbox.submit(messages[1]);
-		// No longer call submitIdAllocation directly
-		// outbox.submitIdAllocation(messages[2]);
 		outbox.submit(messages[3]);
-		idAllocOp = messages[2]; // Set the op to be generated by the mock
-
+		// Configure stub for the flush
+		generateIdAllocationOpStub.onCall(0).returns(messages[2]);
 		outbox.flush();
 
 		// ID Alloc op [2] is generated and prepended to the main batch [0, 1, 3]
